# Removes locking in some of the SPI and I2C APIs to improve performance