--- a/hal/src/stm32f2xx/i2c_hal.c
+++ b/hal/src/stm32f2xx/i2c_hal.c
@@ -906,10 +906,7 @@ void HAL_I2C_Flush_Data(HAL_I2C_Interface i2c, void* reserved)
 
 bool HAL_I2C_Is_Enabled(HAL_I2C_Interface i2c, void* reserved)
 {
-    HAL_I2C_Acquire(i2c, NULL);
-    bool en = i2cMap[i2c]->I2C_Enabled;
-    HAL_I2C_Release(i2c, NULL);
-    return en;
+    return i2cMap[i2c]->I2C_Enabled;
 }
 
 void HAL_I2C_Set_Callback_On_Receive(HAL_I2C_Interface i2c, void (*function)(int), void* reserved)