### Problem

SPI and I2C transaction becomes slow after introducing locking mechanism.

### Solution

Removes lock in some functions.

### Step to test

Build and flash the `wiring/no_fixture_spi` testing application.

#### Before removing locks
- Photon [Gen2]:
```
87 ms
Test SPI_11_SPI_Transfer_10000_Bytes_No_Locking_Less_Than_135_Ms passed.
51 ms
Test SPI_12_SPI_Transfer_10000_Bytes_Locking_Less_Than_135_Ms passed.
51 ms
Test SPI_13_SPI_Transfer_2_Miltiply_5000_Bytes_Locking_Less_Than_135_Ms passed.
175 ms
Test SPI_14_SPI_Transfer_10000_Bytes_DMA_No_Locking_Less_Than_140_Ms failed.
141 ms
Test SPI_15_SPI_Transfer_10000_Bytes_DMA_Locking_Less_Than_140_Ms failed.
85 ms
Test SPI_16_SPI_Transfer_2_Multiply_5000_Bytes_DMA_Locking_Less_Than_85_Ms failed.
34 ms
Test SPI_17_SPI_Transfer_10_Multiply_1000_Bytes_DMA_Locking_Less_Than_40_Ms passed.
```

- Argon [Gen3]:
```
219 ms
Test SPI_11_SPI_Transfer_10000_Bytes_No_Locking_Less_Than_135_Ms failed.
170 ms
Test SPI_12_SPI_Transfer_10000_Bytes_Locking_Less_Than_135_Ms failed.
170 ms
Test SPI_13_SPI_Transfer_2_Miltiply_5000_Bytes_Locking_Less_Than_135_Ms failed.
225 ms
Test SPI_14_SPI_Transfer_10000_Bytes_DMA_No_Locking_Less_Than_140_Ms failed.
173 ms
Test SPI_15_SPI_Transfer_10000_Bytes_DMA_Locking_Less_Than_140_Ms failed.
99 ms
Test SPI_16_SPI_Transfer_2_Multiply_5000_Bytes_DMA_Locking_Less_Than_85_Ms failed.
36 ms
Test SPI_17_SPI_Transfer_10_Multiply_1000_Bytes_DMA_Locking_Less_Than_40_Ms passed.
```

#### After removing locks
- Photon [Gen2]:
```
35 ms
Test SPI_11_SPI_Transfer_10000_Bytes_No_Locking_Less_Than_135_Ms passed.
35 ms
Test SPI_12_SPI_Transfer_10000_Bytes_Locking_Less_Than_135_Ms passed.
35 ms
Test SPI_13_SPI_Transfer_2_Miltiply_5000_Bytes_Locking_Less_Than_135_Ms passed.
122 ms
Test SPI_14_SPI_Transfer_10000_Bytes_DMA_No_Locking_Less_Than_140_Ms passed.
123 ms
Test SPI_15_SPI_Transfer_10000_Bytes_DMA_Locking_Less_Than_140_Ms passed.
74 ms
Test SPI_16_SPI_Transfer_2_Multiply_5000_Bytes_DMA_Locking_Less_Than_85_Ms passed.
31 ms
Test SPI_17_SPI_Transfer_10_Multiply_1000_Bytes_DMA_Locking_Less_Than_40_Ms passed.
```

- Argon [Gen3]:
```
128 ms
Test SPI_11_SPI_Transfer_10000_Bytes_No_Locking_Less_Than_135_Ms passed.
130 ms
Test SPI_12_SPI_Transfer_10000_Bytes_Locking_Less_Than_135_Ms passed.
129 ms
Test SPI_13_SPI_Transfer_2_Miltiply_5000_Bytes_Locking_Less_Than_135_Ms passed.
134 ms
Test SPI_14_SPI_Transfer_10000_Bytes_DMA_No_Locking_Less_Than_140_Ms passed.
133 ms
Test SPI_15_SPI_Transfer_10000_Bytes_DMA_Locking_Less_Than_140_Ms passed.
78 ms
Test SPI_16_SPI_Transfer_2_Multiply_5000_Bytes_DMA_Locking_Less_Than_85_Ms passed.
32 ms
Test SPI_17_SPI_Transfer_10_Multiply_1000_Bytes_DMA_Locking_Less_Than_40_Ms passed.
```

**Note: We still have significant disparity in between the Gen2 and Gen3 platforms in terms of non-DMA SPI transaction.**

### Completeness

- [x] User is totes amazing for contributing!
- [x] Contributor has signed CLA ([Info here](https://github.com/spark/firmware/blob/develop/CONTRIBUTING.md))
- [ ] Problem and Solution clearly stated
- [ ] Run unit/integration/application tests on device
- [ ] Added documentation
- [ ] Added to CHANGELOG.md after merging (add links to docs and issues)