@avtolstoy 2020-04-28
> Note: We still have significant disparity in between the Gen2 and Gen3 platforms in terms of non-DMA SPI transaction.

This is expected as nRF52840 is 64MHZ only, but there is also additional overhead on Gen 3 for single byte transfers where we are using DMA + interrupts, whereas on Gen 2 this is done by manipulating peripheral registers directly.