@avtolstoy 2020-04-28
@XuGuohui I've fixed an issue on Electron and updated the performance tests. Please review.